requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.3.1            # via django
beautifulsoup4==4.9.3     # via django-bootstrap4
django-bootstrap4==2.3.1  # via -r requirements.in
django-fieldsets-with-inlines==0.6  # via -r requirements.in
django==3.1.4             # via -r requirements.in, django-bootstrap4
importlib-metadata==2.1.1  # via django-bootstrap4
pytz==2020.4              # via django
soupsieve==2.1            # via beautifulsoup4
sqlparse==0.4.1           # via django
zipp==3.4.0               # via importlib-metadata
